requirements.txt
# Core dependencies
dspy-ai>=2.4.0
mlflow>=2.8.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API clients
groq>=0.4.0
openai>=1.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0

# Development/testing
pytest>=7.0.0
pytest-cov>=4.0.0
